added more stuff

+++ ExigentDev.DIM.Api/Controllers/StockController.cs
@@ -1,9 +1,8 @@
-using ExigentDev.DIM.Api.Data;
 using ExigentDev.DIM.Api.Dtos.Stock;
+using ExigentDev.DIM.Api.Helpers;
 using ExigentDev.DIM.Api.Interfaces;
 using ExigentDev.DIM.Api.Mappers;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 namespace ExigentDev.DIM.Api.Controllers
 {
@@ -14,62 +13,79 @@ namespace ExigentDev.DIM.Api.Controllers
     private readonly IStockRepository _stockRepo = stockRepo;
 
     [HttpGet]
-    public async Task<IActionResult> GetAll()
+    public async Task<IActionResult> GetAll([FromQuery] QueryObject queryObject)
     {
-      var stocks = await _stockRepo.GetAllAsync();
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
+      var stocks = await _stockRepo.GetAllAsync(queryObject);
 
       var stockDto = stocks.Select(s => s.ToStockDto());
 
       return Ok(stockDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       var stock = await _stockRepo.GetByIdAsync(id);
 
       if (stock == null)
       {
-        return NotFound();
+        return NotFound("Stock not found");
       }
 
       return Ok(stock.ToStockDto());
     }
 
     [HttpPost]
-    public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
+    public async Task<IActionResult> Create([FromBody] CreateStockDto stockDto)
     {
-      var stockModel = stockDto.ToStockFromCreateDTO();
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
+      var stockModel = stockDto.ToStockFromCreateDto();
 
       await _stockRepo.CreateAsync(stockModel);
 
       return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
     }
 
-    [HttpPut("{id}")]
-    public async Task<IActionResult> Update(
-      [FromRoute] int id,
-      [FromBody] UpdateStockRequestDto updateDto
-    )
+    [HttpPut("{id:int}")]
+    public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockDto updateDto)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
       if (stockModel == null)
       {
-        return NotFound();
+        return NotFound("Stock not found");
       }
 
       return Ok(stockModel.ToStockDto());
     }
 
-    [HttpDelete("{id}")]
+    [HttpDelete("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       var stockModel = await _stockRepo.DeleteAsync(id);
 
       if (stockModel == null)
       {
-        return NotFound();
+        return NotFound("Stock not found");
       }
 
       return NoContent();
